Update the issue template
The original issue template was very focused on bugs. I'd really like different
types of issues submitted for bugs than for feature requests, etc so put
instructions into the issue template file and link to the actual templates that
should be used.

Also, add an example of a correctly formatted bug report and feature request for
reference.